Implements support for parameters
With this change it is now possible to use parameters to configure
(void)walker. Only one parameter has been added which is
'voidwalker-theme' that is used to control the default interface theme.

Some of the base code has been redesigned so that it does not depend on
GDB which will simplify testing. Factories are used to create the
correct implementation.