Modifies context to display changed registers using face-special
The context output has been given the ability to check for changes between
the 'previous' context and the current one. This allows it to display modified
registers with the 'special' face which makes it easier to track values.

This change also includes the first (void)walker hook that can be installed to
enable the context to be dumped automatically on GDB hook-stop.