Splits registers into register groups for improved structuring
Each architecture now has the registers divided into groups such as general
purpose, special purpose, floating point etc. It is also possible to add
special register types with the ability to use special register printers.

x86 support has been included. It's now possible to configure the number
of doublewords to show in the stack dump and the number of instructions to
include in the code view.