Restructures (void)walker into three modules: framework, application and backends
(void)walker has been split into three distinct modules. framework provides
the basic framework for the whole application. Application is the
implementation based on the framework an consists of components such as themes,
commands and CPUs. Backends hold the backends that provide data to the
application, currently test and gdb.

A new command is introduced as well, voidwalker dump instructions, which can
be used to display disassemblies using the built-in widgets.