Support for patching binaries in memory
This change includes support for patching binaries when loaded in memory so
that instructions can be modified before execution. It includes a simple
assembler as well as a snippet library which currently only contains a snippet
for patching nops.

On top of this a GDB support command has been included that makes it simple to
add a breakpoint at the .text section start address. This is useful when
setting an initial breakpoint in a stripped binary.